kasasa/map-data-pipeline-flyway/requirements.txt
awscli==1.11.14
boto3==1.7.74
botocore==1.10.74
futures==3.0.5
jmespath==0.9.0
pystache==0.5.4
PyYAML==3.12
six==1.10.0
PyMySQL==0.7.11
backports.tempfile
